requirements.txt
asgiref==3.2.10
beautifulsoup4==4.11.1
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.12
cryptography==36.0.2
defusedxml==0.7.1
Django==3.1
django-allauth==0.50.0
django-bootstrap5==21.3
Faker==13.12.0
idna==3.3
oauthlib==3.2.0
Pillow==7.2.0
psycopg2-binary==2.9.3
pycparser==2.21
PyJWT==2.3.0
python-dateutil==2.8.2
python3-openid==3.2.0
pytz==2020.1
requests==2.27.1
requests-oauthlib==1.3.1
six==1.16.0
soupsieve==2.3.2.post1
sqlparse==0.3.1
urllib3==1.26.9
